refator: Move math and currency utils to dedicated library
