+++ components/contribution-flow/ContributionDetails.js
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedDate, FormattedMessage } from 'react-intl';
 
+import { formatCurrency } from '../../lib/currency-utils';
 import { getNextChargeDate } from '../../lib/date-utils';
-import { formatCurrency } from '../../lib/utils';
 
 import Container from '../Container';
 import { P, Span } from '../Text';
