+++ components/expenses/EditPayExpenseFeesForm.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
-import { getCurrencySymbol } from '../../lib/utils';
+import { getCurrencySymbol } from '../../lib/currency-utils';
 
 import InputField from '../InputField';
