+++ lib/__tests__/currency-utils.test.js
@@ -1,0 +1,9 @@
+import * as utils from '../currency-utils.js';
+
+describe('currency utils lib', () => {
+  test('getCurrencySymbol', () => {
+    expect(utils.getCurrencySymbol('USD')).toEqual('$');
+    expect(utils.getCurrencySymbol('EUR')).toEqual('€');
+    expect(utils.getCurrencySymbol('JPY')).toEqual('¥');
+  });
+});
