+++ lib/__tests__/utils.test.js
@@ -11,10 +11,4 @@ describe('utils lib', () => {
     const arr = [undefined];
     expect(utils.capitalize(arr)).toEqual('');
   });
-
-  test('getCurrencySymbol', () => {
-    expect(utils.getCurrencySymbol('USD')).toEqual('$');
-    expect(utils.getCurrencySymbol('EUR')).toEqual('€');
-    expect(utils.getCurrencySymbol('JPY')).toEqual('¥');
-  });
 });
