+++ lib/currency-utils.js
@@ -1,0 +1,55 @@
+import getSymbolFromCurrency from 'currency-symbol-map';
+
+import { CurrencyPrecision } from './constants/currency-precision';
+
+function getCurrencySymbolFallback(currency) {
+  return Number(0)
+    .toLocaleString('en-US', {
+      style: 'currency',
+      currency,
+      minimumFractionDigits: 0,
+      maximumFractionDigits: 0,
+    })
+    .replace(/(^0\s?)|(\s?0$)/, '');
+}
+
+export function getCurrencySymbol(currency) {
+  return getSymbolFromCurrency(currency) || getCurrencySymbolFallback(currency);
+}
+
+export const getPrecisionFromAmount = amount => {
+  return amount.toString().slice(-2) === '00' ? 0 : CurrencyPrecision.DEFAULT;
+};
+
+function getLocaleFromCurrency(currency) {
+  let locale;
+  switch (currency) {
+    case 'USD':
+      locale = 'en-US';
+      break;
+    case 'EUR':
+      locale = 'en-EU';
+      break;
+    default:
+      locale = currency;
+  }
+  return locale;
+}
+
+export function formatCurrency(amount, currency = 'USD', options = {}) {
+  amount = amount / 100;
+  let minimumFractionDigits = 2;
+  let maximumFractionDigits = 2;
+  if (Object.prototype.hasOwnProperty.call(options, 'minimumFractionDigits')) {
+    minimumFractionDigits = options.minimumFractionDigits;
+  } else if (Object.prototype.hasOwnProperty.call(options, 'precision')) {
+    minimumFractionDigits = options.precision;
+    maximumFractionDigits = options.precision;
+  }
+  return amount.toLocaleString(getLocaleFromCurrency(currency), {
+    style: 'currency',
+    currency,
+    minimumFractionDigits: minimumFractionDigits,
+    maximumFractionDigits: maximumFractionDigits,
+  });
+}
