+++ lib/payment_method_label.js
@@ -16,7 +16,7 @@ import { get, padStart } from 'lodash';
 import moment from 'moment';
 import { defineMessages } from 'react-intl';
 
-import { formatCurrency } from './utils';
+import { formatCurrency } from './currency-utils';
 
 const messages = defineMessages({
   virtualcard: {
