+++ lib/utils.js
@@ -1,7 +1,4 @@
-import getSymbolFromCurrency from 'currency-symbol-map';
 import loadScript from 'load-script';
-
-import { CurrencyPrecision } from './constants/currency-precision';
 
 export function truncate(str, length) {
   if (!str || typeof str !== 'string' || str.length <= length) {
@@ -43,25 +40,6 @@ export const isValidEmail = email => {
   );
 };
 
-function getCurrencySymbolFallback(currency) {
-  return Number(0)
-    .toLocaleString('en-US', {
-      style: 'currency',
-      currency,
-      minimumFractionDigits: 0,
-      maximumFractionDigits: 0,
-    })
-    .replace(/(^0\s?)|(\s?0$)/, '');
-}
-
-export function getCurrencySymbol(currency) {
-  return getSymbolFromCurrency(currency) || getCurrencySymbolFallback(currency);
-}
-
-export const getPrecisionFromAmount = amount => {
-  return amount.toString().slice(-2) === '00' ? 0 : CurrencyPrecision.DEFAULT;
-};
-
 export function parseToBoolean(value) {
   let lowerValue = value;
   // check whether it's string
@@ -82,21 +60,6 @@ export function prettyUrl(url) {
     .replace(/^https?:\/\/(www\.)?/i, '')
     .replace(/\?.+/, '')
     .replace(/\/$/, '');
-}
-
-function getLocaleFromCurrency(currency) {
-  let locale;
-  switch (currency) {
-    case 'USD':
-      locale = 'en-US';
-      break;
-    case 'EUR':
-      locale = 'en-EU';
-      break;
-    default:
-      locale = currency;
-  }
-  return locale;
 }
 
 export function getQueryParams() {
@@ -128,27 +91,6 @@ export function formatDate(date, options = { month: 'long', year: 'numeric' }) {
       return d.toString();
     }
   }
-}
-
-export function formatCurrency(amount, currency = 'USD', options = {}) {
-  amount = amount / 100;
-
-  let minimumFractionDigits = 2;
-  let maximumFractionDigits = 2;
-
-  if (Object.prototype.hasOwnProperty.call(options, 'minimumFractionDigits')) {
-    minimumFractionDigits = options.minimumFractionDigits;
-  } else if (Object.prototype.hasOwnProperty.call(options, 'precision')) {
-    minimumFractionDigits = options.precision;
-    maximumFractionDigits = options.precision;
-  }
-
-  return amount.toLocaleString(getLocaleFromCurrency(currency), {
-    style: 'currency',
-    currency,
-    minimumFractionDigits: minimumFractionDigits,
-    maximumFractionDigits: maximumFractionDigits,
-  });
 }
 
 export const singular = str => {
@@ -240,42 +182,6 @@ export const firstSentence = (str, length) => {
   }
   str = trim(str, length);
   return str;
-};
-
-/*
-* Shortens a number to the abbreviated thousands, millions, billions, etc
-* https://stackoverflow.com/a/40724354
-
-* @param {number} number: value to shorten
-* @returns {string|number}
-
-* @example
-* // return '12.3k'
-* abbreviateNumber(12345, 1)
-*/
-
-const SI_PREFIXES = ['', 'k', 'M', 'G', 'T', 'P', 'E'];
-
-export const abbreviateNumber = (number, precision = 0) => {
-  // what tier? (determines SI prefix)
-  const tier = (Math.log10(number) / 3) | 0;
-
-  const round = value => {
-    return precision === 0 ? Math.round(value) : value.toFixed(precision);
-  };
-
-  // if zero, we don't need a prefix
-  if (tier == 0) {
-    return round(number);
-  }
-
-  // get prefix and determine scale
-  const scale = Math.pow(10, tier * 3);
-
-  // scale the number
-  const scaled = number / scale;
-
-  return round(scaled) + SI_PREFIXES[tier];
 };
 
 export const loadScriptAsync = (url, opts = {}) =>
